+++ module/sheets/item-sheet.mjs
@@ -1,49 +1,98 @@
-export class yuigahamaItemSheet extends foundry.appv1.sheets.ItemSheet {
+import {manageTabs} from "../helpers/common.mjs";
 
-  /** @override */
-  static get defaultOptions() {
-    return foundry.utils.mergeObject(super.defaultOptions, {
-      classes: ["yuigahama", "sheet", "item"],
+const api = foundry.applications.api;
+const sheets = foundry.applications.sheets;
+
+export class yuigahamaItemSheet extends api.HandlebarsApplicationMixin(sheets.ItemSheetV2) {
+
+  static PARTS = {
+    header: {
+      template: `systems/yuigahama/templates/item/item-attribut-sheet.html`
+    },
+  };
+
+  static DEFAULT_OPTIONS = {
+    classes: ["yuigahama", "sheet", "item"],
+    window: {
+      contentClasses: ['standard-form', 'scrollable'],
+      resizable: true,
+    },
+    actions: {},
+    form: {
+      submitOnChange: true,
+      closeOnSubmit: false,
+    },
+    position: {
       width: 550,
       height: 600,
-      tabs: [{ navSelector: ".sheet-tabs", contentSelector: ".sheet-body-item", initial: "description" }]
-    });
+    },
+    dragDrop: [{ dragSelector: "[data-drag]", dropSelector: null }],
+  };
+
+  static TABS = {
+    description : {
+          id: 'description',
+          group: 'primary',
+          title: 'yuigahama.tab.first',
+          icon: '<i class="fa-regular fa-chart-line"></i>'
+        },
+    attributes: {
+          id: 'attributes',
+          group: 'primary',
+          title: 'yuigahama.tab.second',
+          icon: '<i class="fa-solid fa-file-contract"></i>'
+        }
   }
 
-  /** @override */
-  get template() {
-    return `systems/yuigahama/templates/item/item-${this.item.type}-sheet.html`;
+  tabGroups = {
+    primary: 'description'
   }
 
-  /* -------------------------------------------- */
+  getTabs () {
+    const tabs = yuigahamaItemSheet.TABS
 
-  /** @override */
-  getData() {
-    // Retrieve base data structure.
-    const context = super.getData();
-
-    // Use a safe clone of the item data for further operations.
-    const itemData = context.item;
-
-    // Retrieve the roll data for TinyMCE editors.
-    context.rollData = {};
-    let actor = this.object?.parent ?? null;
-    if (actor) {
-      context.rollData = actor.getRollData();
+    for (const tab of Object.values(tabs)) {
+      tab.active = this.tabGroups[tab.group] === tab.id
+      tab.cssClass = tab.active ? 'active' : ''
     }
 
-    // Add the actor's data to context.data for easier access, as well as flags.
-    context.system = itemData.system;
-    context.flags = itemData.flags;
+    return tabs
+  }
+
+  get title() {
+    return `${this.item.name}`;
+  }
+
+  async _prepareContext(options) {
+    const context = await super._prepareContext(options);
+
+    context.item = context.document;
+    // Prepare tabs
+    context.tabs = this.getTabs()
 
     return context;
   }
 
-  /** @override */
-  activateListeners(html) {
-    super.activateListeners(html);
+  async _onRender(context, options) {
+    await super._onRender(context, options);
 
-    if (!this.isEditable) return;
+    //Reminder AppV2 has abandoned jQuery, so, if you still want to use it, you must add the following:
+    const html = $(this.element)
 
+    //Tab active for CSS
+    const tabs = this.getTabs();
+
+    manageTabs(html,tabs,true);
   }
+
+  async _preparePartContext(partId, context, options) {
+    const partContext = await super._preparePartContext(partId, context, options);
+
+    if (partId === this.item.type)
+      partContext.tab = partContext.tabs.description;   // so template can access tab.cssClass
+    else
+      partContext.tab = partContext.tabs[partId];   // so template can access tab.cssClass
+    return partContext;
+  }
+
 }